# BackEndConSwagger/bin/Debug/net6.0/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BackEnd",
    "version": "1.0"
  },
  "paths": {
    "/Parqueadero/Auth/Register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/modUserDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/modUserDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/modUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/modUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/modUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/modUser"
                }
              }
            }
          }
        }
      }
    },
    "/Parqueadero/Auth/Login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/modUserDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/modUserDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/modUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/Parqueadero/Auth/Refresh-token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "modUser": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "passwordHash": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "passwordSalt": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          },
          "tokenCreated": {
            "type": "string",
            "format": "date-time"
          },
          "tokenExpires": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "modUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "oauth2": {
        "type": "apiKey",
        "description": "Standard Authorization header using the Bearer scheme (\"bearer {token}\")",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}